comments everywhere
Did a lot of commenting and shuring up. Mostly trying to make it
legible, and have the logic flow. I think this is in a very useable
state. I need to run some pilot tests to see how well it works. Might
consider saving the tokens at each trial, but whats the point.