renamed curExp to curSess
curSess makes more sense as it includes information about both the
participant and the run. I can format this string to look nice as a data
file and as a plot title. Also made curRec as a way to keep track of
other important recording information. Whether or not I had some special
parameters I wanted to keep track of, etc...