Moved the token iteration from praat to MATLAB
so now the praat batch only makes one token at a time. It takes a little
longer, but its easier for me to control the naming conventions and the
values of the tokens being produced. I think the next step after this
will be to read all these wave files into a matlab file so I dont need
to keep accessing the storage