Overhaul of JND plotting function
I have made the plot bigger, and I am now reporting what the different
trials are. Giving a random Run # (like 5 instead of 1), does not break
the script. Added markers to show the reversals. Added a legend. Fixed
the way the main script counted the number of catch trials. Did lots to
the dimensions of boxes in the plots.