sectioned data with time set up
because the time overwhich the important parts of a trials happen, the
length of each sectioned trial is different, and are not at this time,
aligned by the same tight points. For this reason, each trial's
sectioned mic/head will have its own time vector associated with it. It
will become important that this time vector accompany each frequency
analysis, so that it can be determined where in the trial, this lines
up.