better control of which trials are saved
Keeping a better record overall of which trials are saved post processing in both the time domain processing and in the frequency domain processing. More information is provided to the researcher about which trial is thrown away, and more appropriate variable names are used to make make the progression of analyses more legible.  Changed the audioPostProcessing function quite a bit, most importantly, how it uses previously created variables. It also doesnt filter the audio files themselves but instead parses which trials should be saved, and the parsing is done outside of the function.