renaming a function; pairing dowin its utility
the function SaveRawData was a left over back from when I was saving my
data trial by trial. The most importat thing it did was load my raw
Audapter data from the buffer. It also saved my wav files, but without
regard to the types of trials being run. Now this script will only save
the wav files and catch statements have been placed in the main
recording window so that the act is done appropriately. Also pulling the
Audapter raw data off the buffer is done in line again. Maybe eventually
it needs to be its own function, but for now it can live here.