Improved sorting and cat methods
Before there was a lot of hard coded methods for organizing just two conditions, and unknown number of runs for that condition. Now cat allows for any number of runs to be sorted, but I think this will probably still break if there is an uneven number of runs between subjects. There is still a hard coded mess for how many conditions we, have and its hard coded to 2 at the moment. Hopefully this will become cleaner in the future. Will require more clever coding.