REMOVING RESULTS FOLDER
This repo is being transformed. Its original goal was to handle presentation, analysis, and results for my dissertation project, but now I am working to make it simply for presentation and analysis. From now on the results from my, and others experiments, will be in the repo labeled 'dfResults'. This transition will be marked in each repo as a release.