Fixed weird doubling glitch
MY GOD did this take a while to fix. I had initially thought that loading the file once and using it on each offline trial would have worked, but it would seem that Audapter/MATLAB overwrote the Mic_frames variable after each trial, essentially rewriting that mic name. Reopening it in the for loops makes things pretty again. I could probably split this up into loading the file, and making the cells, we will give that a go later.