applying dScale param
Was perviously ignoring this param, but using it now. This allows the output to be scaled by  the multiplier. We these settings, and settings in CueMix and on Behringer, the headphones will play Produced voice at roughly 5dB louder than microphone recording. Also setting formantbypass to true in order to remove unnecessary processing steps.