updating these config generators
I am mostly proud of like 43 in LarynxPosFenConfig. This tells pooled analysis to find the last number of the run name and use that to find the collar position we used, since we didnt write that down specifically.

Made a second config file that tests AudFB instead of collar location.