No more output variables
Now saving results in a results matlab data structure at a dirs dependent on 'subject' and 'run'.

If recordNewData is set to 0, the it loads the data that was requested, and does analysis on it.

Also fixed script so that live result window pops up before the start of the first trial.