updating the SetAudFB to always have correct loudness
I was hard setting the headphone gain, but that was lazy. Now the dB value is set in the individual experiments, and the gain decimal value is calculated based on the input value. I am also setting up the setAudFB to be more congrous with the new version of Audapter, so that the other branches can take advantage of these changes.