small change to offline analysis
This is confirmed to work simply through changing the dfSetAudFB...I am going to make a few other small file changes so things are less confusing. This works with using the default methods for pitch-shifting from the original Audapter changes, but can also use the peaks and valleys versions as well.