Removing PraatCheck from SFPerturb
We used to take a look at the recorded practice trials in praat right after they were recorded to see, not only how were the recordings, but how well did my app compare to the Gold Standard of Praat. Since we are relatively comfortable with the output of the coding app, we dont need praat at this step anymore. Not gone forever. Can br brought back if needed.