First pass at Over Haul of the NIDAQ Analysis
putting sensor dynamics in its own structure for better safe keeping. This will allow for interative adding to any or all sensor dynamics without making the script much longer. In this current form, the rise time, and Plateau start values are wrong. Will fix this in the next commit. I will also need to change the other plotting functions that reply on some of these values later. For now I test the script and all fo the mean, SE functions are working, and are being saved, and are being plotted. The guts of what needed to be changed are working. Good job team.