Updating the JND plotting function
there was an included variable called tN which did not make any sense. I have since called in selectOpt since it identifies the options that the participant can make, when they select an answer on a given trial.

Also a legacy portion of this script saved the jpg files with a suffix denoting the type of JND run. I've removed it since, for the current moment we are not testing different JND types and its a bit misleading