Nearly final updates to the dfAnalysisJND methods
These are some nearly updates for analyzing the raw JND data. Overall I can now deduce the JND score for a given run and save it separately. This was a bit overkill, but I thought it important to keep the JND score calcuation methods, and aggregate pooled result methods separate. This means that if we want to calculate the score (even for a given run for a given subject), we can do that pretty painlessly. Similarly if we want to rerun the stats, without needing to rerun the individual JND calcuation for each individual participant/run, that can happen to. The changes to the pooledresult function will happen next.