reorg of RespVar
these variables were creeping towards becoming a point where it was hard to keep track of them. Also, I needed to make some analogous for the AudPert schema.  so here we are. The analysis pipeline is ready, but need to go back and check the pooled analysis and plotting functions