Addressing Voice Onset error
Ultimately I made the auditory pert trials a little too short. Looks like, at times when voice onset is late (like close to 1s), the trial is not long enough to be captured for analysis. There is not much to be done at this point, since the pert happens at voice onset, and I need to align everything there. May need to readdress the recordings/participants if I throw away too many trials.