Updating RunPooledAnalysis for new res structure types
need to update this pooled analysis script to account for the new result structure type that is chugging through. Also updating the InflationResponse script to match what is in the AnalysisAudio Script. I think I need both since, because I need to individual run versions, and the all the trial versions.