Adjusting PreProcessing step to save trials that are too short
I think ultimately I should have made these trials longer, Too often in these analyses, the participant has enough useful pert information, but the trial is too short (sometimes by a hair), and it totally gets tossed. I am fixing this to pad all trials that are too short, so that it will be saved and used in analysis. I might update this in the future so that there are more check conditions, but ultimately there are more behaviroal responses in the onset period that I can still gleam some information from without throwing away the whole trial.