Quality of life improvements to Video App
This app certainly does the parsing, but the hope is that it will do so much more soon. In this update, I have added features that let you choose the working directory to find the raw data, load different subjects and runs, and scroll through different frames of a given trial. For the most part, There are some dumb bugs hidden throughout, but they will be caught in the coming iterations. I hope I dont spend much more time on the functionality of the viewing, and I can spend more time completing the actions of measuring the pixels