Adjusting limits problems
It was high time that cleaned up my limits calculating script. The short answer for why things went crazy..is because I had some things hard coded that expected the audio lengths to be a set length. I think it should be able to withstand more changes in the future. really cleaned up the hard coded stuff and made things more simple for plotting