Video analysis app now takes advantage of line objects
In the previous iteration, this app would reply on global properties to keep track of curPoints or the fidLines because I had at most 2 points and 1 line. Since I am looking to add more points and more lines I am making my functions a bit less restrictive on global parameters, and organizing the function calls to flow more logically. I want to be able to add cases for app properties, instead of executing changes for all lines all at the same time. The next step will be to add more lines, and make sure that it all saves properly.