f0 Analysis organization: fV
the function(s) that handles the frequency variables for f0 calculation were disorganized and confusing. I have made an effort to be more transparent about which variables are used where and when. the function setFreqAnalVar, which is a subfunction of AnalysisAudio, is now called from within the actual subfunction signalFrequencyAnalysis, instead of in the outer function. data structures handle important variables, and I have updated Calcf0Praat to handle the structure variables