MH Align Class: Cleaning up env and triggers
The ultimate goal of this class is to time-align the M and H recordings from Audpater, and then also time-align the M Audapter and M NIDAQ recordings. In the latter case, the M Audapter recording is literally shaved so that the trigger time points are correct in time-space for the recording. The trigger values are not changed...the recording is shifted so that the triggers align to the correct section of the recording. That said, I do create a new set of triggers/time points representing the veritable trigger points in the raw signal (pre-alignment). Dont use these except for simple intermediate visualization of the raw signals