Removing NIDAQ Plotting functions: DAQAll and DAQPresMic
Removing the functions 'drawDAQAll' and 'drawDAQPresMic'. 'drawDAQAll' was very unorganized and had roughly 2x as many channels to view as are necessary. If we end up adding more channels in the future, it should be pretty simple to make a better version of this. 'drawDAQPresMic essentially created individual trial versions of what 'drawAllPertTrialMicf0' does now. The latter is much cleaner and effecienet, making the former redundant. There are comments added through out this commit, as well as removal of function calls for the aforementioned removed drawing functions.