Zero-Padding during the sectioning of trials
In some very rare case, due to the shift of alignment from the trigger onset to inflation onset, I would attempt an offset sectioning beyond the bounds of the signal. For this reason, I am zeropadding the signal to avoid errors involved in those fringe cases, and causing the loop to break. This should not affect the results at all, but should make the analysis go smoothly.