New Perturbatron signal, slight adjustment
I wrote some extra code here for the addition of the valve switcher for the new perturbatrons. The guts of dfMakePertSignal are still pretty relevant, but before I pass this into the master branch I am getting rid of the function outputs. I had only gone through the effort of setting this up for the dfDiagnostics_Sensors and not the other scripts that call this function. For this reason, I want to scale back these features until I know that they will actually be implemented in the future.